Split stageInstances types
