+++ packages/types/src/discord/guilds.ts
@@ -1,18 +1,11 @@
 /** Types for: https://discord.com/developers/docs/resources/guild */
 
-import type {
-  DiscordAvatarDecorationData,
-  DiscordRole,
-  DiscordStageInstance,
-  DiscordSticker,
-  DiscordUser,
-  DiscordVoiceState,
-  OAuth2Scope,
-} from '../discord.js'
+import type { DiscordAvatarDecorationData, DiscordRole, DiscordSticker, DiscordUser, DiscordVoiceState, OAuth2Scope } from '../discord.js'
 import type { DiscordChannel } from './channels.js'
 import type { DiscordEmoji } from './emojis.js'
 import type { DiscordPresenceUpdate } from './gateway.js'
 import type { DiscordSoundboardSound } from './soundboards.js'
+import type { DiscordStageInstance } from './stageInstances.js'
 
 /** https://discord.com/developers/docs/resources/guild#guild-object-guild-structure */
 export interface DiscordGuild {
